sep class
